Move imported classes to `com.novoda.noplayer.external.exoplayer`
