+++ core/src/main/java/com/novoda/noplayer/external/exoplayer/text/webvtt/CssParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.novoda.noplayer.exoplayer.text.webvtt;
+package com.novoda.noplayer.external.exoplayer.text.webvtt;
 
 import android.text.TextUtils;
