+++ core/src/main/java/com/novoda/noplayer/external/exoplayer/util/ColorParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.novoda.noplayer.exoplayer.util;
+package com.novoda.noplayer.external.exoplayer.util;
 
 import android.text.TextUtils;
